Improving implementation of mounted profile files.

+++ bin/reset.sh
@@ -18,6 +18,7 @@ source conf/docker.env
 [ -f "${PWD}/.container" ] && source ${PWD}/.container
 WORKING_DIRECTORY=
 
+count() { echo $#; }
 check_lockfile() { echo "$(cat ${1})"; }
 get_container_by_id() { docker ps -aq --filter id="$1"; }
 get_container_by_name() { docker ps -aq --filter name="$1"; }
@@ -36,6 +37,7 @@ cleanup_dotfiles() {
             [ "${item_basename}" = ".gnupg" ] && targeted+=(${WORKING_DIRECTORY}/${item_basename})
             [ "${item_basename}" = ".kube" ] && targeted+=(${WORKING_DIRECTORY}/${item_basename})
             [ "${item_basename}" = ".password-store" ] && targeted+=(${WORKING_DIRECTORY}/${item_basename})
+            [ "${item_basename}" = ".profile.d" ] && targeted+=(${WORKING_DIRECTORY}/${item_basename})
         done
     fi
 
@@ -51,24 +53,18 @@ cleanup_dotfiles() {
 }
 
 clean_lockfile() {
-    printf "\033[93m>\033[0m Checking for lockfile: %s\n" "$1"
     if [ -f "${PWD}/${1}" ]; then
-        printf "\033[93m>\033[0m Removing ...\n"
+        printf "\033[93m>\033[0m Removing '%s' ...\n" "$1"
         rm -f "${PWD}/${1}"
-    else
-        printf "\033[91m>\033[0m Not Found. %s\n" "${NOTHING_MSG}"
     fi
 }
 
 quick_clean() {
     if [ -n "$1" ]; then
         local targeted="$(get_container_by_name ${1})"
-        printf "\033[93m>\033[0m Checking for container by name: %s (%s)\n" "${1}" "${targeted}"
         if [ -n "$targeted" ]; then
-            printf "\033[93m>\033[0m Removing ...\n"
+            printf "\033[93m>\033[0m Removing container %s (%s) ...\n" "${1}" "${targeted}"
             echo $targeted | xargs docker rm -f
-        else
-            printf "\033[91m>\033[0m Not Found. %s\n" "${NOTHING_MSG}"
         fi
     fi
 
@@ -80,7 +76,7 @@ full_clean() {
     cleanup_dotfiles $(ls -d ${WORKING_DIRECTORY}/.*)
     if [ -f "${_MOUNT_CACHE_FILE}" ]; then
         RANDOMSTR="$(check_lockfile ${_MOUNT_CACHE_FILE})"
-        printf "\033[93m>\033[0m Removing docker volume: %s\n" "${RANDOMSTR}"
+        printf "\033[93m>\033[0m Removing docker volume '%s' ...\n" "${RANDOMSTR}"
         docker volume rm "${RANDOMSTR}"
         clean_lockfile "${_MOUNT_CACHE_FILE}"
     fi
@@ -106,7 +102,9 @@ printf "\033[92mCleanup Mode: \033[92;1m%s\033[0m\n" "${mode}"
 quick_clean "${CONTAINER_NAME}"
 
 if [ "$mode" = "full" ]; then
-    [ -d "${PWD}/mount/dotfiles/${DOCKER_USER}" ] && full_clean "${PWD}/mount/dotfiles/${DOCKER_USER}"
-    printf "\033[93m>\033[0m Removing '%s' ...\n" "${PWD}/mount/addons"
-    rm -rf "${PWD}/mount/addons"
+    [ -d "${PWD}/mount/home/${DOCKER_USER}" ] && full_clean "${PWD}/mount/home/${DOCKER_USER}"
+    if [ -d "${PWD}/mount/addons" ]; then
+        printf "\033[93m>\033[0m Removing '%s' ...\n" "${PWD}/mount/addons"
+        rm -rf "${PWD}/mount/addons"
+    fi
 fi
